+++ apps/wasm/mw.c
@@ -8,7 +8,7 @@
 const char *version()
 {
     void * p = malloc(32);
-    strcpy(p, "m - 0.0.36");
+    strcpy(p, "m - 0.0.37");
     return p;
 }